backend/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.8
dataclasses-json==0.6.7
en-core-web-sm==3.7.1
fastapi==0.114.1
filelock==3.16.0
frozenlist==1.4.1
fsspec==2024.9.0
greenlet==3.1.0
grpcio==1.66.1
grpcio-tools==1.66.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.24.7
hyperframe==6.0.1
idna==3.8
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.1
langchain-huggingface==0.1.0
langchain-qdrant==0.1.3
langchain-text-splitters==0.3.0
langcodes==3.4.0
langdetect==1.0.9
langid==1.1.6
langsmith==0.1.119
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
orjson==3.10.7
packaging==24.1
pillow==10.4.0
pip==24.2
portalocker==2.10.1
preshed==3.0.9
protobuf==5.28.1
pydantic==2.9.2
pydantic_core==2.23.4
pydantic-settings==2.5.2
Pygments==2.18.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pypdf==4.3.1
python-dotenv==1.0.1
pywin32==306
PyYAML==6.0.2
qdrant-client==1.11.1
regex==2024.9.11
requests==2.32.3
rich==13.8.1
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.1.0
setuptools==72.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.34
srsly==2.4.8
starlette==0.38.5
sympy==1.13.2
tenacity==8.5.0
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.1
tqdm==4.66.5
transformers==4.44.2
typer==0.12.5
typing_extensions==4.12.2
typing-inspect==0.9.0
urllib3==2.2.3
uvicorn==0.30.6
wasabi==1.1.3
watchfiles==0.24.0
weasel==0.4.1
websockets==13.0.1
wheel==0.44.0
wrapt==1.16.0
yarl==1.11.1
